feat: add code for retrieving signed url resources
